2014/03/05 11:11

+++ laser/ButtonManager.cpp
@@ -16,8 +16,8 @@ ButtonManager::ButtonManager()
 
 void ButtonManager::update(sf::RenderWindow& window)
 {
-	mouseBounds_.left = sf::Mouse::getPosition(window).x;
-	mouseBounds_.top = sf::Mouse::getPosition(window).y;
+	mouseBounds_.left = (float)(sf::Mouse::getPosition(window).x);
+	mouseBounds_.top = (float)(sf::Mouse::getPosition(window).y);
 
 	if(sf::Mouse::isButtonPressed(sf::Mouse::Left))
 	{
